requirements.txt
streamlit
pandas
matplotlib
google-api-python-client
python-dotenv
scikit-learn
joblib
langdetect
